refactor: 방 종료 검증 로직 위치 Migrator에서 RoomFacade.resetRoom() 으로 수정 #268

+++ backend/src/main/java/ddangkong/facade/room/RoomFacade.java
@@ -4,6 +4,7 @@ import ddangkong.domain.balance.content.BalanceContent;
 import ddangkong.domain.room.Room;
 import ddangkong.domain.room.member.Member;
 import ddangkong.domain.room.member.RoomMembers;
+import ddangkong.exception.room.NotFinishedRoomException;
 import ddangkong.facade.room.dto.InitialRoomResponse;
 import ddangkong.facade.room.dto.RoomInfoResponse;
 import ddangkong.facade.room.dto.RoomJoinResponse;
@@ -14,7 +15,6 @@ import ddangkong.facade.room.member.dto.MemberResponse;
 import ddangkong.service.balance.content.BalanceContentService;
 import ddangkong.service.room.RoomService;
 import ddangkong.service.room.balance.roomcontent.RoomContentService;
-import ddangkong.service.room.balance.roomvote.RoomBalanceVoteService;
 import ddangkong.service.room.balance.roomvote.RoomMigrator;
 import ddangkong.service.room.member.MemberService;
 import java.time.LocalDateTime;
@@ -35,8 +35,6 @@ public class RoomFacade {
     private final RoomContentService roomContentService;
 
     private final BalanceContentService balanceContentService;
-
-    private final RoomBalanceVoteService roomBalanceVoteService;
 
     private final RoomMigrator roomMigrator;
 
@@ -109,9 +107,15 @@ public class RoomFacade {
 
     @Transactional
     public void resetRoom(Long roomId) {
-        Room room = roomService.reset(roomId);
+        Room room = roomService.getRoom(roomId);
+
+        if (!room.isGameFinish()) {
+            throw new NotFinishedRoomException();
+        }
+
+        room.reset();
         roomContentService.deleteRoomContents(room);
-        roomMigrator.migrateFinishedRoom(room);
+        roomMigrator.migrateRoom(room);
     }
 
     @Transactional
